Sept 16 commit

+++ src/functions/CalcAvgFreq.jl
@@ -1,12 +1,8 @@
 
-
 ##############################################
 #		FUNCTION CALC AVG FREQUENCIES        #
 ##############################################
-function CalcMedian(ii::Int64, no_columns::Int64, no_param::Int64, x_dist::Matrix{Float64})
-#function CalcMedian(p::CalcStatArg)
-#    @unpack ii, no_columns, no_param, x_dist = p
-
+function CalcMedian(ii::Int, no_columns::Int, no_param::Int, x_dist::Matrix{Float64})
     # Step 1: Quantitative trait medians (ignoring NaN values)
 	qt_medians = mapslices(median ∘ skipmissing, x_dist[x_dist[:, 1] .== ii, 2:no_param+1], dims=1)
 	# Step 2: Discrete trait frequencies
@@ -17,10 +13,7 @@ function CalcMedian(ii::Int64, no_columns::Int64, no_param::Int64, x_dist::Matri
 	return median_freqs
 end
 
-function CalcMean(ii::Int64, no_columns::Int64, no_param::Int64, x_dist::Matrix{Float64})
-#function CalcMean(p::CalcStatArg)
-#    @unpack ii, no_columns, no_param, x_dist = p
-
+function CalcMean(ii::Int, no_columns::Int, no_param::Int, x_dist::Matrix{Float64})
 	# Step 1: Quantitative trait medians (ignoring NaN values)
 	qt_means = mapslices(mean ∘ skipmissing, x_dist[x_dist[:, 1] .== ii, 2:no_param+1], dims=1)
 	# Step 2: Discrete trait frequencies
@@ -31,14 +24,12 @@ function CalcMean(ii::Int64, no_columns::Int64, no_param::Int64, x_dist::Matrix{
 	return mean_freqs
 end
 
-function CalcVar(ii::Int64, no_param::Int64, x_dist::Matrix{Float64})
-#function CalcVar(p::CalcStatArg)
-#    @unpack ii, no_columns, no_param, x_dist = p
-
+function CalcVar(ii::Int, no_param::Int, x_dist::Matrix{Float64})
     temp_size = size(x_dist[x_dist[:,1] .== ii,2:no_param+1])
 	if temp_size[1] == 1 #has only one row = only one individual
 		trait_var = Matrix{Float64}(fill(0, temp_size[1], length(2:no_param+1)))
-	else
+	else 
+		# get variance
 		trait_var = mapslices(var ∘ skipmissing, x_dist[x_dist[:,1] .== ii,2:no_param+1],dims=1)
 	end
 	return trait_var
